+++ NEWS.md
@@ -1,5 +1,13 @@
 # lintr (development version)
 
+## New and improved features
+
+### New linters
+
+* `unnecessary_anonymous_function_linter()`: detect unnecessary anonymous functions, e.g.
+  `lapply(x, function(xi) sum(xi))` can be `lapply(x, sum)` and `purrr::map(x, ~quantile(.x, 0.75, na.rm = TRUE))`
+  can be `purrr::map(x, quantile, 0.75, na.rm = TRUE)`. Naming `probs = 0.75` can further improve readability.
+
 # lintr 3.0.1
 
 * Skip multi-byte tests in non UTF-8 locales (#1504)